+++ src/parse_execution.rs
@@ -57,7 +57,7 @@ use fish_common::{ScopeGuard, escape, help_section, truncate_at_nul};
 use fish_widestring::WExt as _;
 use libc::{EXIT_SUCCESS, STDERR_FILENO, STDOUT_FILENO, c_int};
 use nix::errno::Errno;
-use std::{io::ErrorKind, rc::Rc, sync::Arc};
+use std::{rc::Rc, sync::Arc};
 
 /// An eval_result represents evaluation errors including wildcards which failed to match, syntax
 /// errors, or other expansion errors. It also tracks when evaluation was skipped due to signal
@@ -264,47 +264,78 @@ impl ExecutionContext {
         ctx: &mut OperationContext<'_>,
         cmd: &wstr,
         statement: &ast::DecoratedStatement,
-        err: std::io::Error,
+        errno: Errno,
     ) -> EndExecutionReason {
-        // We couldn't find the specified command. This is a non-fatal error. We want to set the exit
-        // status to 127, which is the standard number used by other shells like bash and zsh.
+        // We couldn't find the specified command or it wasn't executable. This is a non-fatal error.
+        // Set the exit status to 126 or 127, which are the standard numbers used by other shells like bash and zsh.
 
-        if err.kind() != ErrorKind::NotFound {
-            // TODO: We currently handle all errors here the same,
-            // but this mainly applies to EACCES. We could also feasibly get:
-            // ELOOP
-            // ENAMETOOLONG
-            if matches!(Errno::try_from(err), Ok(Errno::ENOTDIR)) {
-                // If the original command did not include a "/", assume we found it via $PATH.
-                let src = self.node_source(&statement.command);
-                if !src.contains('/') {
-                    return report_error!(
-                        self,
-                        ctx,
-                        STATUS_NOT_EXECUTABLE,
-                        &statement.command,
-                        "Unknown command. A component of '%s' is not a directory. Check your $PATH.",
-                        cmd
-                    );
-                } else {
-                    return report_error!(
-                        self,
-                        ctx,
-                        STATUS_NOT_EXECUTABLE,
-                        &statement.command,
-                        "Unknown command. A component of '%s' is not a directory.",
-                        cmd
-                    );
+        if errno != Errno::ENOENT {
+            // STATUS_NOT_EXECUTABLE (126)
+            let error_msg = match errno {
+                // error_msg = reason
+                Errno::ENOMEM => localize!("enomem" = "Out of memory."),
+                Errno::EMFILE => localize!("emfile" = "Too many open files in this process."),
+                Errno::ENFILE => localize!("enfile" = "Too many open files on the system."),
+                _ => {
+                    // error_msg = "Unknown command <cmd> " + reason
+                    let reason = match errno {
+                        Errno::ETXTBSY => {
+                            localize!("etxtbsy" = "File is currently open for writing")
+                        }
+                        Errno::ELOOP => {
+                            localize!("eloop" = "Too many levels of symbolic links. Maybe a loop?")
+                        }
+                        Errno::ENAMETOOLONG => localize!("enametoolong" = "File name is too long."),
+                        Errno::EISDIR => localize!("eisdir" = "File is a directory."),
+                        Errno::ENOTDIR => {
+                            // If the original command did not include a "/", assume we found it via $PATH.
+                            let src = self.node_source(&statement.command);
+                            if !src.contains('/') {
+                                localize!(
+                                    "enotdir-on-PATH" = "One of the parent files is not a directory. Check your $PATH."
+                                )
+                            } else {
+                                localize!("enotdir" = "One of the parent files is not a directory.")
+                            }
+                        }
+                        Errno::EPERM => localize!(
+                            "eperm" = "No permission. Either suid/sgid is forbidden or you lack capabilities."
+                        ),
+                        // if the file is not a regular file, the errno was reassigned to ENOEXEC (not EACCES); see path_get_path_core in path.rs
+                        Errno::EACCES => localize!(
+                            "eacces-modified" = "Missing execute permissions for the file or one of its parent directories."
+                        ),
+                        // real ENOEXECs caught elsewhere; only the synthetic ENOEXEC from path_get_path_core will reach here
+                        Errno::ENOEXEC => {
+                            localize!("enoexec-modified" = "Files of this type cannot be executed.")
+                        }
+                        // EIO, EOVERFLOW, etc are rare but possible
+                        _ => {
+                            let err = match errno {
+                                Errno::UnknownErrno => "unknown".to_owned(),
+                                _ => format!("{:?}", errno), // POSIX error code name (e.g. 'ENOMEM')
+                            };
+                            localize!(
+                                "command-not-found-catchall" =
+                                    "The file exists but cannot be executed. Error: { $err }.",
+                                err = err
+                            )
+                        }
+                    };
+                    // now construct `error_msg` with `reason`
+                    localize!(
+                        "command-not-found" = "Unknown command '{ $cmd }': { $reason }",
+                        cmd = cmd,
+                        reason = reason
+                    )
                 }
-            }
-
-            return report_error!(
+            };
+            return report_error_formatted!(
                 self,
                 ctx,
                 STATUS_NOT_EXECUTABLE,
                 &statement.command,
-                "Unknown command. '%s' exists but is not an executable file.",
-                cmd
+                WString::from_str(&error_msg)
             );
         }
 
@@ -782,7 +813,7 @@ impl ExecutionContext {
                             &external_cmd.path
                         },
                         statement,
-                        std::io::Error::from(external_cmd.err.unwrap()),
+                        external_cmd.err.unwrap(),
                     );
                 }
             }